Fix a few bugs with progress bars
The filled percentage on the progress bars was capped, except for the ones under "Current Jobs." This could result in the filled part of these bars overflowing the bar container (and even the entire screen) for a single tick while rapidly advancing through professions within a category (using auto-promote).

After that, the levels shown in progress bars were not formatted the same way as job and skill levels in their respective tasks, leading to numbers showing up through without commas (or equivalent), which is suboptimal in most cultures. For example: "Concentration Lv1234" versus "Concentration Lv1,234"

Finally, there was a space between "Lv" and the number for the offline time progress bar in the stats page, which is inconsistent with the progress bars in the sidebar which have no such space. This was fixed by rewriting the relevant HTML code to remove whitespace between "Lv" and the element that holds the level number.